create,delete,show add

+++ app/redux/features/books/books.ts
@@ -1,28 +1,32 @@
+import { Book } from "@/types/books";
 import { apiSlice } from "../../api/apiSlice";
-
-export interface Book {
-    id?: number;
-    title: string;
-    author: string;
-    availableCopies: number;
-}
 
 export const bookApi = apiSlice.injectEndpoints({
     endpoints: (builder) => ({
-        getBooks: builder.query<Book[], void>({
+        getBooks: builder.query<{ data: Book[] }, undefined>({
             query: () => "books",
         }),
-        getBookById: builder.query<Book, number>({
+        getBookById: builder.query<Book, string>({
             query: (id) => `books/${id}`,
         }),
+
         addBook: builder.mutation<Book, Partial<Book>>({
-            query: (book) => ({
-                url: "books",
-                method: "POST",
-                body: book,
-            }),
+            query: (book) => {
+                // console.log(book);
+                return {
+                    url: "books",
+                    method: "POST",
+                    body: book,
+                };
+            },
+
+            //  ({
+            //     url: "books",
+            //     method: "POST",
+            //     body: book,
+            // }),
         }),
-        updateBook: builder.mutation<Book, { id: number; data: Partial<Book> }>(
+        updateBook: builder.mutation<Book, { id: string; data: Partial<Book> }>(
             {
                 query: ({ id, data }) => ({
                     url: `books/${id}`,
@@ -31,11 +35,18 @@ export const bookApi = apiSlice.injectEndpoints({
                 }),
             }
         ),
-        deleteBook: builder.mutation<void, number>({
-            query: (id) => ({
-                url: `books/${id}`,
-                method: "DELETE",
-            }),
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        deleteBook: builder.mutation<Book, string>({
+            query: (id) => {
+                if (!id) {
+                    console.log(id);
+                    throw new Error("Books ID is Required");
+                }
+                return {
+                    url: `books/${id}`,
+                    method: "DELETE",
+                };
+            },
         }),
     }),
 });
@@ -47,3 +58,57 @@ export const {
     useUpdateBookMutation,
     useDeleteBookMutation,
 } = bookApi;
+
+// export const bookApi = apiSlice.injectEndpoints({
+//     endpoints: (builder) => ({
+//         getBooks: builder.query<Book[], void>({
+//             query: () => "books",
+//             transformResponse: (response: any[]) =>
+//                 response.map((book) => ({
+//                     id: book.bookId, // Map `bookId` to `id`
+//                     title: book.title,
+//                     author: book.author ?? "Unknown Author", // Handle optional fields
+//                     availableCopies: book.availableCopies,
+//                 })),
+//         }),
+//         getBookById: builder.query<Book, string>({
+//             query: (id) => `books/${id}`,
+//             transformResponse: (response: any) => ({
+//                 id: response.bookId, // Map `bookId` to `id`
+//                 title: response.title,
+//                 author: response.author ?? "Unknown Author",
+//                 availableCopies: response.availableCopies,
+//             }),
+//         }),
+//         addBook: builder.mutation<Book, Partial<Book>>({
+//             query: (book) => ({
+//                 url: "books",
+//                 method: "POST",
+//                 body: {
+//                     bookId: book.id, // Map `id` back to `bookId` if needed
+//                     title: book.title,
+//                     author: book.author,
+//                     availableCopies: book.availableCopies,
+//                 },
+//             }),
+//         }),
+//         updateBook: builder.mutation<Book, { id: string; data: Partial<Book> }>(
+//             {
+//                 query: ({ id, data }) => ({
+//                     url: `books/${id}`,
+//                     method: "PUT",
+//                     body: {
+//                         bookId: id,
+//                         ...data,
+//                     },
+//                 }),
+//             }
+//         ),
+//         deleteBook: builder.mutation<void, string>({
+//             query: (id) => ({
+//                 url: `books/${id}`,
+//                 method: "DELETE",
+//             }),
+//         }),
+//     }),
+// });
